Rework conformer, remove some code.
